backend/pyproject.toml
[project]
name = "twitterfake"
version = "0.1.0"
description = ""
authors = [
    {name = "Erick",email = "ericklemos2001@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "faker (>=40.8.0,<41.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
